Default parameters; thermostat adjustment; deprioritize longest hachures
contour & hachure density now get a default value that can be adjusted. Will help new users, and especially later on if I make this into a proper tool vs. a script.

Thermostat fix: Adjusted thresholds for too_short / too_long to prevent rapid cycling on/off of hachures that are near-parallel and keep crossing the threshold by a tiny bit.

Now when a split is too_short, we choose randomly which line to clip. This helps aesthetics, as the longest line can look odd sometimes.